load earthquake data
